fix(codegen,objfile,tests): Windows x86_64 native assembler pipeline fixes
- Fix CoffWriter cross-section symbol resolution: rodata symbols (.LC_str_*)
  were emitted as undefined in the COFF symbol table, causing LNK2001 errors.
  Process rodata symbols first and redirect text relocations to the defined
  rodata entries, matching the MachO writer's dedup approach.

- Fix X64BinaryEncoder missing runtime symbol mapping: external calls used
  raw IL names (Viper.Terminal.PrintStr) instead of C runtime names
  (rt_print_str). Added mapRuntimeSymbol() matching AArch64 pattern.

- Fix CrossLayerArithTests hang: missing dispatchChild() guard caused
  infinite process recursion on Windows via ProcessIsolation CreateProcess
  self-relaunch.

- Add -lshell32 to Windows linker command for DragQueryFile/DragAcceptFiles.

- Apply same cross-section symbol fix to ElfWriter (preventative).

+++ src/runtime/oop/rt_lazy.c
@@ -32,6 +32,7 @@
 
 #include "rt_lazy.h"
 #include "rt_object.h"
+#include "rt_platform.h"
 #include "rt_string.h"
 
 #include <stdlib.h>
@@ -43,7 +44,7 @@
 typedef enum { VALUE_PTR = 0, VALUE_STR = 1, VALUE_I64 = 2 } ValueType;
 
 typedef struct {
-    int8_t evaluated;
+    int evaluated; /* widened from int8_t for MSVC _Generic atomic compat */
     ValueType value_type;
     void *(*supplier)(void);
